# locker-server/docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.ErrorResponse:
    properties:
      error:
        example: invalid credentials
        type: string
    type: object
  handlers.HoldSuccessResponse:
    properties:
      expires_in:
        example: 5 minutes
        type: string
      locker:
        $ref: '#/definitions/handlers.LockerResponse'
      message:
        example: locker held successfully
        type: string
    type: object
  handlers.ListLockersResponse:
    properties:
      available_count:
        example: 45
        type: integer
      lockers:
        items:
          $ref: '#/definitions/handlers.LockerResponse'
        type: array
    type: object
  handlers.LockerResponse:
    properties:
      location_id:
        type: string
      locker_id:
        type: integer
      owner:
        description: null 가능
        type: string
    type: object
...
  handlers.MyLockerResponse:
    properties:
      locker:
        $ref: '#/definitions/handlers.LockerResponse'
    type: object
...
  handlers.SimpleSuccessResponse:
    properties:
      message:
        example: operation completed successfully
        type: string
    type: object
info:
  contact: {}
  description: 사물함 선착순 예약 시스템의 백엔드 API 문서
  title: Locker Reservation API
  version: "1.0"
paths:
...
  /lockers:
    get:
      consumes:
      - application/json
      description: 모든 사물함 목록과 사용 가능한 사물함 수를 반환합니다
      parameters:
      - default: Bearer
        description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 사물함 목록과 사용 가능 수
          schema:
            $ref: '#/definitions/handlers.ListLockersResponse'
        "401":
          description: 인증 필요
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: 서버 오류
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 사물함 목록 조회
      tags:
      - lockers
  /lockers/{id}/confirm:
    post:
      consumes:
      - application/json
      description: 선점한 사물함을 확정합니다 (실제 소유권 획득). hold 상태에서 confirmed 상태로 전환되며, 사물함의
        소유자로 등록됩니다.
      parameters:
      - default: Bearer
        description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 사물함 ID (선점한 사물함)
        example: 101
        in: path
        maximum: 999
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 확정 완료
          schema:
            $ref: '#/definitions/handlers.SimpleSuccessResponse'
        "400":
          description: 잘못된 요청 - 유효하지 않은 사물함 ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: 인증 필요 - JWT 토큰이 없거나 유효하지 않음
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: 선점이 만료되었거나 없음 - hold 상태가 아니거나 5분이 경과함
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: 서버 오류 - 데이터베이스 트랜잭션 실패
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 사물함 확정
      tags:
      - lockers
  /lockers/{id}/hold:
    post:
      consumes:
      - application/json
      description: 특정 사물함을 선점합니다 (5분간 예약). Redis와 DB를 통해 동시성 제어를 하며, 성공 시 사물함 정보를
        반환합니다.
      parameters:
      - default: Bearer
        description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 사물함 ID (1-999 범위)
        example: 101
        in: path
        maximum: 999
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: 선점 성공 - 사물함 정보 포함
          schema:
            $ref: '#/definitions/handlers.HoldSuccessResponse'
        "400":
          description: 잘못된 요청 - 유효하지 않은 사물함 ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: 인증 필요 - JWT 토큰이 없거나 유효하지 않음
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: 이미 선점됨 - 다른 사용자가 이미 선점했거나 본인이 이미 선점한 상태
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "503":
          description: 서비스 일시 불가 - Redis 서버 장애
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 사물함 선점
      tags:
      - lockers
  /lockers/{id}/release:
    post:
      consumes:
      - application/json
      description: 확정된 사물함을 해제합니다 (소유권 포기). confirmed 상태에서 cancelled 상태로 전환되며, 사물함이
        다시 사용 가능해집니다.
      parameters:
      - default: Bearer
        description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 사물함 ID (소유한 사물함)
        example: 101
        in: path
        maximum: 999
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 해제 완료
          schema:
            $ref: '#/definitions/handlers.SimpleSuccessResponse'
        "400":
          description: 잘못된 요청 - 유효하지 않은 사물함 ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: 인증 필요 - JWT 토큰이 없거나 유효하지 않음
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: 사물함을 찾을 수 없음 - 소유하지 않은 사물함
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: 서버 오류 - 데이터베이스 트랜잭션 실패
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 사물함 해제
      tags:
      - lockers
  /lockers/me:
    get:
      consumes:
      - application/json
      description: 현재 로그인한 사용자가 소유한 사물함 정보를 반환합니다. 소유한 사물함이 없으면 locker가 null로 반환됩니다.
      parameters:
      - default: Bearer
        description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 내 사물함 정보 (없으면 locker가 null)
          schema:
            $ref: '#/definitions/handlers.MyLockerResponse'
        "401":
          description: 인증 필요 - JWT 토큰이 없거나 유효하지 않음
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: 서버 오류 - 데이터베이스 조회 실패
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 내 사물함 조회
      tags:
      - lockers
securityDefinitions:
  BearerAuth:
    description: Bearer {access_token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"